Reactivate image linking.
This reactivates the nightly linking up of new images/instances to one another.
This was disabled to not interfere with clearing out all of the images recors
(which had incorrect IDs).

This also changes "what images to consider for linking" to any modified, instead
of any created. This is because the recreation of those image records that *had*
correct IDs (which will now be "undeleted") must also again lead to linking.
This act of "undeletion" does not however reset the created timestamp, which is
why we (at least for the moment) must select on "newly changed" images, and not
just "new" images. This part should (perhaps) be reverted when all the images
are in place again.