feat(search2): Allow escaping of one-char wildcard(?)
When ? only appears as the last character of a word (or words) we don't
treat it as a wildcard. It is unlikely that the user intent is to use a
wildcard. This makes pasted titles work as queries e.g. "who is who?".

If ? is used in any other position, wildcards are enabled. Including trailing.
examples: "blom??uist", "wom??", "?ome?".

Add the possibility to force a single trailing ? to be interpreted as a
wildcard by escaping it: "who\?". (or maybe rather "unescaping" it)

TODO?
We might want to handle this in a different way in the lexing step.