feat(search2): Restrictions on category value
Add support for restrictions on property value. That is, we can now tell
if category: X in a query corresponds to _categoryByCollection. find/identify/none.
So that it for example works having category find/identify/none facets
in different filtering phases in ES. And we only filter in the relevant field.

Property can now be a "value restricted term" which can then be "narrowed" based on its value.
category + saogf:Poesi -> identifyCategory (= _categoryByCollection.identify)
The "narrowed" version can also be created directly by name, e.g. identifyCategory. This is used in facet definitions.